separated utility functions into two files to avoid circular dependency, got rid of labels parameter for EventReportItem and ChartReportItem
